Adding auth.js and passport.js and updating the index.js endpoints
If applied, this commit will add authentication logic that will authenticate login request. Update the endpoints in the Index.js file to incorporate the authentication process and allow users to make request. Also adding Passport authentication for Node and express that allows certain means of authentication and authorization multiple times throughout the application.